Add new preset to Tado to enable geofencing mode
Add new Auto preset mode to enable Tado to be set to auto geofencing
mode.  The existing Home and Away presets switch Tado into manual
geofencing mode and there was previously no way to restore it to auto
mode.
Note 1: All three preset modes apply to the Tado home holistically, not
just the climate component used to select the preset.  To make this
clearer, three new sensors have also been added to display the state
of the Tado home
Note 2: Auto mode is only supported if the Auto Assist skill is enabled
in the owner's Tado home. Various checks have been added to ensure the
Tado supports auto geofencing and if it is not supported, the preset is
not listed in the preset modes listed
Note 3: The PyTado library has been updating to set auto mode